SetSizes Fix/Songear Comments Fix
SetSize function was calling getProductPrice too early, before the size
value was inserted into the DOM and the getProductPrice wasn’t finding
anything to use for the size to ajax.

Singer change was based on a couple of typos: 1) capital %S instead of
lowercase %s, and sprinf instead of sprintf.